refactor: extract process management into internal/process package (#335)
Move process lifecycle files (proctracker, resourcemonitor, envtag,
signal handling, child PID discovery) from internal/eval to a new
internal/process package. This reduces the eval package's scope to
evaluation orchestration only.

- Export NotifyShutdownSignals and FindChildCopilotPIDs for
  cross-package use
- Update engine.go, copilot.go, client.go, engine_test.go imports
