+++ hyoka/internal/eval/client.go
@@ -5,6 +5,7 @@ import (
 	"log/slog"
 
 	copilot "github.com/github/copilot-sdk/go"
+	"github.com/ronniegeraghty/hyoka/hyoka/internal/process"
 )
 
 // BuildBaseClientOpts returns a shared set of Copilot SDK client options
@@ -13,7 +14,7 @@ import (
 // to ensure consistent configuration.
 func BuildBaseClientOpts() *copilot.ClientOptions {
 	opts := &copilot.ClientOptions{
-		Env: HyokaBaseEnv(),
+		Env: process.HyokaBaseEnv(),
 	}
 	if slog.Default().Enabled(context.Background(), slog.LevelDebug) {
 		opts.LogLevel = "debug"
