+++ hyoka/internal/process/proctracker.go
@@ -1,4 +1,4 @@
-package eval
+package process
 
 import (
 	"log/slog"
