+++ hyoka/internal/process/proctracker_test.go
@@ -1,4 +1,4 @@
-package eval
+package process
 
 import (
 	"os"
