+++ hyoka/internal/process/resourcemonitor.go
@@ -1,4 +1,4 @@
-package eval
+package process
 
 import (
 	"fmt"
