+++ hyoka/internal/process/resourcemonitor_test.go
@@ -1,4 +1,4 @@
-package eval
+package process
 
 import (
 	"testing"
