+++ hyoka/internal/process/signal_windows.go
@@ -1,15 +1,15 @@
 //go:build windows
 
-package eval
+package process
 
 import (
 	"os"
 	"os/signal"
 )
 
-// notifyShutdownSignals registers ch to receive os.Interrupt (Ctrl+C).
+// NotifyShutdownSignals registers ch to receive os.Interrupt (Ctrl+C).
 // Windows does not have SIGTERM; the closest equivalent is os.Interrupt
 // which is delivered for both Ctrl+C and Ctrl+Break.
-func notifyShutdownSignals(ch chan<- os.Signal) {
+func NotifyShutdownSignals(ch chan<- os.Signal) {
 	signal.Notify(ch, os.Interrupt)
 }
